Structurize and refactor
