+++ babel.config.js
@@ -1,3 +1,13 @@
 module.exports = {
   presets: ['module:@react-native/babel-preset'],
+  plugins: [
+    [
+      'module-resolver',
+      {
+        alias: {
+          src: './src',
+        },
+      },
+    ],
+  ],
 };
