+++ src/api/index.ts
@@ -10,37 +10,6 @@ export const apiClient = Axios.create({
   baseURL: BASE_URL,
 });
 
-export enum Status {
-  ALIVE = 'Alive',
-  DEAD = 'Dead',
-  UNKNOWN = 'unknown',
-}
-
-type Origin = {
-  name: string;
-  url: string;
-};
-
-type Location = {
-  name: string;
-  url: string;
-};
-
-export type Character = {
-  id: string;
-  name: string;
-  status: Status;
-  species: string;
-  type: string;
-  gender: string;
-  origin: Origin;
-  location: Location;
-  image: string;
-  episode: string[];
-  url: string;
-  created: string;
-};
-
 export type CharactersResponse = {
   results: Character[];
 };
