+++ src/api/types.d.ts
@@ -1,0 +1,30 @@
+enum Status {
+  ALIVE = 'Alive',
+  DEAD = 'Dead',
+  UNKNOWN = 'unknown',
+}
+
+type Origin = {
+  name: string;
+  url: string;
+};
+
+type Location = {
+  name: string;
+  url: string;
+};
+
+type Character = {
+  id: string;
+  name: string;
+  status: Status;
+  species: string;
+  type: string;
+  gender: string;
+  origin: Origin;
+  location: Location;
+  image: string;
+  episode: string[];
+  url: string;
+  created: string;
+};
