+++ src/navigation/types.d.ts
@@ -1,0 +1,4 @@
+type RootStackParamList = {
+  CharactersList: undefined;
+  CharacterDetail: { id: Character['id'] };
+};
