+++ src/screens/CharactersListScreen/components/Photo.tsx
