+++ src/screens/CharactersListScreen/components/index.ts
@@ -1,0 +1,7 @@
+import Photo from './Photo';
+import CharacterItem from './CharacterItem';
+
+export {
+  Photo,
+  CharacterItem,
+};
