+++ src/screens/CharactersListScreen/index.ts
@@ -1,0 +1,1 @@
+export { default as CharactersListScreen } from './CharactersListScreen';
