+++ src/screens/index.ts
@@ -1,0 +1,7 @@
+import { CharactersListScreen } from './CharactersListScreen';
+import { CharacterDetailScreen } from './CharacterDetailScreen';
+
+export {
+  CharactersListScreen,
+  CharacterDetailScreen,
+};
